feat(dot-manager): simplify script, centralized logging

+++ .config/dot-manager/install/fonts.sh
@@ -3,26 +3,21 @@
 source "$DOT_MANAGER_DIR/helper.sh"
 
 install_fonts() {
-  local cmd="$1"
+  cd /tmp || exit 1
+  print_step "Installing Nerd Fonts"
+  [ -d /tmp/nerdfonts ] && rm -rf /tmp/nerdfonts
 
-  cd /tmp
-  case "$cmd" in
-  "update")
-    log "info" "Installing Nerd Fonts..."
-    if [ -d /tmp/nerdfonts ]; then
-      log "info" "Removing old Nerd Fonts directory."
-      rm -rf /tmp/nerdfonts
-    fi
-
-    wget "https://github.com/ryanoasis/nerd-fonts/releases/download/v3.4.0/NerdFontsSymbolsOnly.zip" &&
-      unzip NerdFontsSymbolsOnly.zip -d /tmp/nerdfonts &&
-      mkdir -p "$HOME/.local/share/fonts" &&
-      cp /tmp/nerdfonts/*.ttf "$HOME/.local/share/fonts" &&
-      rm -rf /tmp/nerdfonts NerdFontsSymbolsOnly.zip
-    fc-cache -fv
-    log "info" "Nerd Fonts installed successfully."
-    ;;
-  esac
+  if wget -q "https://github.com/ryanoasis/nerd-fonts/releases/download/v3.4.0/NerdFontsSymbolsOnly.zip" &&
+    unzip -oq NerdFontsSymbolsOnly.zip -d /tmp/nerdfonts &&
+    mkdir -p "$HOME/.local/share/fonts" &&
+    cp /tmp/nerdfonts/*.ttf "$HOME/.local/share/fonts" &&
+    rm -rf /tmp/nerdfonts NerdFontsSymbolsOnly.zip; then
+    fc-cache -fv >>"$DOT_MANAGER_LOG" 2>&1
+    log "success" "Nerd Fonts installed"
+  else
+    log "error" "Nerd Fonts installation failed"
+    return 1
+  fi
 }
 
-install_fonts "$@"
+[ "$1" = "update" ] && install_fonts
